+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.7.0 - Unreleased
 - Don't trigger command to highlight region on selection modified
+- [Goto Analysis Finding #55](https://github.com/pedrorgirardi/Pep/issues/55)
 
 ## 0.6.0 - 2022-06-17
 - New command **Pep: Outline**